Added authentication routers
